issue-1559: [Disk Manager] Move snapshot transfer test to snapshot service test
We should rename snapshot transfer to snapshot creation.
