+++ cloud/disk_manager/internal/pkg/facade/ya.make
@@ -27,8 +27,7 @@ RECURSE_FOR_TESTS(
     filesystem_scrubbing_nemesis_test
     filesystem_service_test
     filesystem_snapshot_service_test
-    filesystem_snapshot_transfer_test
-    filesystem_snapshot_transfer_nemesis_test
+    filesystem_snapshot_service_nemesis_test
     placement_group_service_nemesis_test
     placement_group_service_test
     private_service_nemesis_test
